requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
aiohttp==3.8.3
    # via openai
aiosignal==1.3.1
    # via aiohttp
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via
    #   channels
    #   django
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==21.4.0
    # via
    #   aiohttp
    #   jsonschema
billiard==3.6.4.0
    # via celery
blis==0.7.9
    # via thinc
catalogue==2.0.8
    # via
    #   spacy
    #   srsly
    #   thinc
celery[redis]==5.2.7
    # via -r requirements/requirements.in
celery-progress==0.1.2
    # via -r requirements/requirements.in
certifi==2022.5.18.1
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
channels==4.0.0
    # via -r requirements/requirements.in
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   nltk
    #   typer
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
confection==0.0.4
    # via thinc
cryptography==37.0.2
    # via pyjwt
cymem==2.0.7
    # via
    #   preshed
    #   spacy
    #   thinc
dataclasses-json==0.5.7
    # via gpt-index
debugpy==1.6.5
    # via -r requirements/requirements.in
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via redis
dj-stripe==2.7.2
    # via -r requirements/requirements.in
django==3.2.16
    # via
    #   -r requirements/requirements.in
    #   channels
    #   dj-stripe
    #   django-allauth
    #   django-allauth-2fa
    #   django-anymail
    #   django-hijack
    #   django-otp
    #   djangorestframework
    #   drf-spectacular
    #   jsonfield
django-allauth==0.50.0
    # via
    #   -r requirements/requirements.in
    #   django-allauth-2fa
django-allauth-2fa==0.10.0
    # via -r requirements/requirements.in
django-analytical==3.1.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==8.6
    # via -r requirements/requirements.in
django-environ==0.9.0
    # via -r requirements/requirements.in
django-hijack==3.2.1
    # via -r requirements/requirements.in
django-otp==1.1.3
    # via django-allauth-2fa
django-waffle==3.0.0
    # via -r requirements/requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
drf-spectacular==0.22.1
    # via -r requirements/requirements.in
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0.tar.gz
    # via -r requirements/requirements.in
filelock==3.9.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
gpt-index==0.2.8
    # via -r requirements/requirements.in
huggingface-hub==0.11.1
    # via transformers
idna==3.3
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via drf-spectacular
jinja2==3.1.2
    # via
    #   spacy
    #   torch
joblib==1.2.0
    # via nltk
jsonfield==3.1.0
    # via dj-stripe
jsonschema==4.6.0
    # via drf-spectacular
kombu==5.2.4
    # via celery
langchain==0.0.65
    # via gpt-index
langcodes==3.3.0
    # via spacy
lxml==4.9.2
    # via python-docx
markupsafe==2.1.2
    # via jinja2
marshmallow==3.19.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
marshmallow-enum==1.5.1
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
murmurhash==1.0.9
    # via
    #   preshed
    #   spacy
    #   thinc
mypy-extensions==0.4.3
    # via typing-inspect
networkx==3.1
    # via torch
nltk==3.8.1
    # via gpt-index
numpy==1.24.1
    # via
    #   blis
    #   gpt-index
    #   langchain
    #   pandas
    #   spacy
    #   thinc
    #   transformers
oauthlib==3.2.0
    # via requests-oauthlib
openai==0.27.0
    # via
    #   -r requirements/requirements.in
    #   gpt-index
packaging==21.3
    # via
    #   huggingface-hub
    #   marshmallow
    #   redis
    #   spacy
    #   thinc
    #   transformers
pandas==1.5.2
    # via
    #   -r requirements/requirements.in
    #   gpt-index
pathy==0.10.1
    # via spacy
plotly==5.13.1
    # via -r requirements/requirements.in
preshed==3.0.8
    # via
    #   spacy
    #   thinc
prompt-toolkit==3.0.29
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pydantic==1.10.4
    # via
    #   confection
    #   langchain
    #   spacy
    #   thinc
pyjwt[crypto]==2.4.0
    # via django-allauth
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via pandas
python-docx==0.8.11
    # via -r requirements/requirements.in
python3-openid==3.2.0
    # via django-allauth
pytz==2022.1
    # via
    #   celery
    #   django
    #   djangorestframework
    #   pandas
pyyaml==6.0
    # via
    #   drf-spectacular
    #   huggingface-hub
    #   langchain
    #   transformers
qrcode==7.3.1
    # via django-allauth-2fa
redis==4.3.1
    # via celery
regex==2022.10.31
    # via
    #   nltk
    #   tiktoken
    #   transformers
requests==2.27.1
    # via
    #   django-allauth
    #   django-anymail
    #   huggingface-hub
    #   langchain
    #   openai
    #   requests-oauthlib
    #   spacy
    #   stripe
    #   tiktoken
    #   transformers
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.5.12
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
smart-open==6.3.0
    # via
    #   pathy
    #   spacy
spacy==3.5.0
    # via
    #   -r requirements/requirements.in
    #   en-core-web-md
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.4
    # via spacy
sqlalchemy==1.4.46
    # via langchain
sqlparse==0.4.2
    # via django
srsly==2.4.6
    # via
    #   confection
    #   spacy
    #   thinc
stripe==3.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
sympy==1.12
    # via torch
tenacity==8.1.0
    # via
    #   gpt-index
    #   plotly
thinc==8.1.8
    # via spacy
tiktoken==0.4.0
    # via gpt-index
tokenizers==0.13.2
    # via transformers
torch==2.0.1
    # via -r requirements/requirements.in
tqdm==4.64.1
    # via
    #   huggingface-hub
    #   nltk
    #   openai
    #   spacy
    #   transformers
transformers==4.25.1
    # via gpt-index
typer==0.7.0
    # via
    #   pathy
    #   spacy
typing-extensions==4.4.0
    # via
    #   huggingface-hub
    #   pydantic
    #   torch
    #   typing-inspect
typing-inspect==0.8.0
    # via dataclasses-json
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.9
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wasabi==1.1.1
    # via
    #   spacy
    #   thinc
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.14.1
    # via deprecated
yarl==1.8.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
